pyproject.toml
[project]
name = "costguard"
version = "0.1.0"
description = "GCP FinOps AI Agent using Google's ADK to analyze cost, detect budget overruns, and provide real-time cloud savings recommendations."
authors = [
  { name = "Jitendra Gupta", email = "jitu028@gmail.com" }
]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
google-cloud-aiplatform = { version = "^1.93.0", extras = [
    "adk",
    "agent-engines",
    "websockets = <12.0",
    "packaging = <24.0"

] }

python = "^3.9"
google-genai = "^1.9.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
google-adk = "^1.0.0"

[tool.poetry.group.dev]
optional = true


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
google-adk = { version = "^1.0.0", extras = ["eval"] }
pytest-asyncio = "^0.26.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"


